manager-flask/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyJWT==2.9.0
psycopg[binary]==3.2.10
gunicorn==23.0.0
